Refactor code to resubmit transactions on RPC error
- As of now, if transaction submition failed due to "invalid json RPC
Response", transaction payload will be marked to be resubmitted to the
transactions queue to be later picked up by one of the workers and
submitted again to the network.

- Payloads marked with resubmit=true, will not be signed nor they will
be picking up a new nonce. Reason is that failure was caused by network
issue not a nonce issue.

- In the case that network issue persisted (more than 500
millsecond),Overseer will pick up on that event and will stop the world
to be resumed later once network connectivity is restored.